Set up Backend, Prisma schema
