+++ backend/prisma/schema.prisma
@@ -9,3 +9,63 @@ datasource db {
   provider = "mongodb"
   url      = env("DATABASE_URL")
 }
+
+model Trip {
+  id String @id @default(auto()) @map("_id") @db.ObjectId
+  location TripLocation
+  startDate DateTime @default(now())
+  endDate DateTime
+  pendingInvitations String[]
+  participants User[] @relation(fields: [participantsID], references: [id])
+  activities String[]
+  participantsID String[] @db.ObjectId
+  image Photo
+}
+
+type TripLocation {
+  city String
+  state String
+  latitude Float
+  longitude Float
+}
+
+model User {
+  id String @id @default(auto()) @map("_id") @db.ObjectId
+  userName String
+  password String
+  friends   User[]   @relation("UserFriends", fields: [friendIDs], references: [id])
+  friendIDs String[] @db.ObjectId
+  addedBy   User[]   @relation("UserFriends", fields: [addedByIDs], references: [id])
+  addedByIDs String[] @db.ObjectId
+  tripID String[] @db.ObjectId
+  trips Trip[] @relation(fields: [tripID], references: [id])
+  firstName String
+  lastName String
+  dateOfBirth DateTime 
+  avatar Photo
+}
+
+model Activity {
+  id String @id @default(auto()) @map("_id") @db.ObjectId
+  name String
+  description String
+  startTime DateTime
+  endTime DateTime
+  location ActivityLocation
+  notes String
+  netUpvotes Int @default(0)
+  isOnCalendar Boolean @default(false)
+  category String[]
+}
+
+type ActivityLocation {
+  address String
+  city String
+  state String
+}
+
+type Photo {
+  height Int
+  width  Int
+  url    String
+}
